[AC-4835] Broke up ListView since Users are meta-weird
